Bound exitPubSubMode fwdDone wait and use WriteInt64 for sub counts
- Add timeout to exitPubSubMode fwdDone wait, matching cleanup() pattern
- Switch all subscription count replies from WriteInt to WriteInt64
  for RESP integer consistency and 32-bit safety

+++ proxy/proxy.go
@@ -263,7 +263,7 @@ func (p *ProxyServer) startPubSubSession(conn redcon.Conn, cmdName string, args 
 		} else {
 			session.patternSet[ch] = struct{}{}
 		}
-		dconn.WriteInt(session.subCount())
+		dconn.WriteInt64(int64(session.subCount()))
 	}
 	if err := dconn.Flush(); err != nil {
 		dconn.Close()
